Implement new system of Achievements. Fix some UI bugs. Add rotation button IG.

+++ ANDROID_PROJET/src/io/brothers/sgm/User/SGMSavedData.java
@@ -7,5 +7,5 @@ import java.util.Map;
  * Created by Simon on 17/03/2015.
  */
 public class SGMSavedData {
-    public Map<String, Integer> data = new HashMap<>();
+    public Map<String, Float> data = new HashMap<>();
 }
